Made both left glue and right glue in GlueStrategy use essentially the same algorithm
Previously I had been using a combination of Array.Fill and a for loop for left glue. Implementing right glue showed I only need the for loop. We can now see the two algorithms are the same but with the operators reversed i.e. (`gluedAt + hint` and `gluedAt - hint` etc).