Implemented an indexer for Line
I had a choice of exception to throw here. Arrays throw `IndexOutOfRangeException` but `List<T>` throws `ArgumentOutOfRangeException`. There didn't seem to be a compelling reason to pick one over the other, but at least with the latter we can easily put the actual value on the exception.

This comes with a `Length` property because we're anticipating most consumers are going to want to use this with a `for` loop.

We deliberately have not inclued a setter because our type is supposed to be immutable.