Changed CellState to be readonly record struct
By encoding the boolean and character representation of the cell state into an actual type we could remove repeated `switch` expressions from `Line`. This also gives us the ability to control construction of the type so we don't have to worry about things such as a naughty consumer casting an `int` to the enum, even if that value isn't a member of the enum.